communiFind, meetMate 분리완료
