+++ FinalProject/src/main/resources/templates/meetmate/meetMateList.html
@@ -12,7 +12,7 @@
 <script type="text/javascript" src="http://code.jquery.com/jquery-latest.js"></script>
 <link rel="icon" type="image/png" th:href="@{http://localhost:8080/image/로고.png}"/>
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-<script src="http://localhost:8080/js/gatchi/meetMateListHG.js"></script>
+<script src="http://localhost:8080/js/gatchi/list.js"></script>
 <script src="js/includeHTML.js"></script>
 
 
@@ -24,11 +24,10 @@
 
   <div class="meetTitle">
     <!-- <img src="/image/gatchiImage/meetmateList.png"/> -->
-    <!-- <img src="/image/gatchiImage/meetmateLogo.png"/> -->
     <h1>Meet Mate는 Meet + Mate의 합성어입니다.</h1>
   </div>
   <div class="meetTitle_Sub">
-    <p>동일한 취미를 가진 사람들의 가치있는 하루를 위한 원데이 모임입니다.</p>
+    <p>동일한 관심사를 가진 사람들의 가치있는 하루를 위한 원데이 모임입니다.</p>
   </div>
 
   <div class="body7">
